% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestIC.R
\name{bestIC}
\alias{bestIC}
\title{Function to identify the best-matching component}
\usage{
bestIC(true, estimated)
}
\arguments{
\item{true}{the reference data}

\item{estimated}{the estimated data, from which the best-matching components will be identified}
}
\value{
The resulting dataset gives the best-matching component for each reference component
}
\description{
This function aims to identify the best-matching component with the reference component
}
\examples{
# generate data
set.seed(1234)
dat1 = mvtnorm::rmvnorm(n = 100, mean = rep(0, 2), sigma = matrix(c(1,0.8,0.8,1), nrow = 2, ncol = 2)) # sigma: covariance matrix
dat2 = mvtnorm::rmvnorm(n = 100, mean = rep(0, 2), sigma = matrix(c(1,0.5,0.5,1), nrow = 2, ncol = 2))
dat3 = matrix(rnorm(300, 0, 1), nrow = 100, ncol = 3)
true = cbind(dat1[,1],dat2[,1]); estimated = cbind(dat1[,2],dat2[,2],dat3)
colnames(true) = paste0("comp", 1:2); colnames(estimated) = paste0("comp", 1:5)
# find the best-matching component
cor(true, estimated)
bestIC(true, estimated)
}
