% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_complete.R
\name{data_complete}
\alias{data_complete}
\title{Function to keep the complete cases}
\usage{
data_complete(data_list)
}
\arguments{
\item{data_list}{input modality data (as a list)}
}
\value{
The resulting dataset is the data containing the complete cases (as a list)
}
\description{
This function keeps the complete cases from the input modality data
}
\examples{
# generate data
dat = data_generate_MAR(nsubj = 100, seed = 1234, n_miss = 20)
# missing data (stored in lists)
dat_miss = NULL; dat_miss[[1]] = dat$Y1_miss; dat_miss[[2]] = dat$Y2_miss
# keep complete cases (stored in lists)
dat_complete = data_complete(dat_miss)
}
