% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate_MAR.R
\name{data_generate_MAR}
\alias{data_generate_MAR}
\title{Data generation function for MAR}
\usage{
data_generate_MAR(
  nsubj = 100,
  nvoxel1 = 1000,
  nvoxel2 = 3000,
  true_ncomp = 2,
  seed = 1234,
  n_miss = 5,
  es_C1H1 = 0.5,
  es_C2H2 = 0.5
)
}
\arguments{
\item{nsubj}{number of subjects}

\item{nvoxel1}{number of voxels in modality 1}

\item{nvoxel2}{number of voxels in modality 2}

\item{true_ncomp}{number of true component}

\item{seed}{seed}

\item{n_miss}{number of missing subjects}

\item{es_C1H1}{effect size for C1 and H1}

\item{es_C2H2}{effect size for C2 and H2}
}
\value{
The resulting dataset has lists including the no-missing data (Y1, Y2), order of subjects who were implemented missing (subj1_miss, subj2_miss), missing data (Y1_miss, Y2_miss), and other intermediate parameters during the data generation process.
}
\description{
This function generates the simulation data (2 modalities) under the setting of MAR
}
\examples{
dat = data_generate_MAR(nsubj = 100, seed = 1234, n_miss = 5); names(dat)
}
