% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate_MCAR.R
\name{data_generate_MCAR}
\alias{data_generate_MCAR}
\title{Data generation function for MCAR}
\usage{
data_generate_MCAR(
  nsubj = 100,
  nvoxel1 = 1000,
  nvoxel2 = 3000,
  true_ncomp = 2,
  seed = 1234,
  n_miss = 5
)
}
\arguments{
\item{nsubj}{number of subjects}

\item{nvoxel1}{number of voxels in modality 1}

\item{nvoxel2}{number of voxels in modality 2}

\item{true_ncomp}{number of true component}

\item{seed}{seed}

\item{n_miss}{number of missing subjects}
}
\value{
The resulting dataset has lists including the no-missing data (Y1, Y2), order of subjects who were implemented missing (subj1_miss, subj2_miss), missing data (Y1_miss, Y2_miss), and other intermediate parameters during the data generation process.
}
\description{
This function generates the simulation data (2 modalities) under the setting of MCAR
}
\examples{
dat = data_generate_MCAR(nsubj = 100, seed = 1234, n_miss = 5); names(dat)
}
