% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{data_standardization}
\alias{data_standardization}
\title{Data standardization function}
\usage{
data_standardization(raw_data)
}
\arguments{
\item{raw_data}{input data}
}
\value{
The resulting dataset is the standardized data
}
\description{
This function standardizes the input data
}
\examples{
dat = data_generate_MCAR(nsubj = 100, seed = 1234, n_miss = 5); head(dat$Y1[,1:5])
Y1_std = data_standardization(dat$Y1); head(Y1_std[,1:5])
}
