% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filica.R
\name{filica}
\alias{filica}
\title{Function to run FI-LICA}
\usage{
filica(
  ncomp = 5,
  flica_niter = 1500,
  n = 20,
  flica_niter2 = 1000,
  rescale = TRUE
)
}
\arguments{
\item{ncomp}{number of components to be used in LICA}

\item{flica_niter}{number of iterations to be used in FI-LICA's Step 1 (completer case analysis)}

\item{n}{number of FI-LICA updates}

\item{flica_niter2}{number of iterations to be used in FI-LICA's Step 2}

\item{rescale}{rescale H and XW (default is TRUE)}
}
\value{
The resulting dataset includes the results estimated from FI-LICA, convergence measure for H and for XW, and the dF history in LICA
}
\description{
This function runs FI-LICA. Note: current function only supports the analysis on 2 modalities.
}
\examples{
# FI-LICA
#re_filica = filica(ncomp = 5, flica_niter = 1500, n = 20, flica_niter2 = 1000, rescale = TRUE)
}
