% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filica_step1.R
\name{filica_step1}
\alias{filica_step1}
\title{Function to run Step 1 in FI-LICA}
\usage{
filica_step1(re, rescale = TRUE)
}
\arguments{
\item{re}{the saved results from completer case analysis using LICA}

\item{rescale}{rescale H and XW (default is TRUE)}
}
\value{
The resulting list includes the estimated H, XW, and the last dF from Step 1 in FI-LICA
}
\description{
This function runs Step 1 in FI-LICA.
}
\examples{
# run Step 1 in FI-LICA
#re_step1 = filica_step1(re = re_completer, rescale = TRUE) ## need subj_miss, mod_std, mod_std_complete
}
