% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filica_step2.R
\name{filica_step2}
\alias{filica_step2}
\title{Function to run Step 2 in FI-LICA}
\usage{
filica_step2(n, ncomp, flica_niter2, H_prev, XW_prev, rescale = TRUE)
}
\arguments{
\item{n}{number of FI-LICA updates}

\item{ncomp}{number of components to be used in LICA}

\item{flica_niter2}{number of iterations to be used in LICA}

\item{H_prev}{the estimated H from Step 1}

\item{XW_prev}{the estimated XW from Step 1}

\item{rescale}{rescale H and XW (default is TRUE)}
}
\value{
The resulting dataset includes the results estimated from Step 2 in FI-LICA, convergence measure for H and for XW, and the dF history in LICA
}
\description{
This function runs Step 2 in FI-LICA. Note: current function only supports the analysis on 2 or 3 modalities.
}
\examples{
# results from Step 1 in FI-LICA
#re_step1 = filica_step1(re = re_completer, rescale = TRUE)
# run Step 2 in FI-LICA
#re_step2 = filica_step2(n = 20, ncomp = 5, flica_niter2 = 1000, H_prev = re_step1$H, XW_prev = re_step1$XW, rescale = TRUE)
}
