% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flica_completer.R
\name{flica_completer}
\alias{flica_completer}
\title{Function to run LICA on complete cases}
\usage{
flica_completer(ncomp, flica_niter)
}
\arguments{
\item{ncomp}{number of components to be used in LICA}

\item{flica_niter}{number of iterations to be used in LICA}
}
\value{
The resulting dataset is the results estimated from LICA
}
\description{
This function calls the matlab function to run LICA on complete cases
}
\examples{
# Save data at "./MATLAB_code/flica/data_std.mat"
#writeMat("./MATLAB_code/flica/data_std.mat",
#         mod1_std_cmplt = mod1_std_cmplt, mod2_std_cmplt = mod2_std_cmplt # complete cases from standardized missing data
#         )
# Run LICA on completers
#re_completer = flica_completer(ncomp = 5, flica_niter = 1500)
}
