% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flica_nomiss.R
\name{flica_nomiss}
\alias{flica_nomiss}
\title{Function to run LICA on no-missing data}
\usage{
flica_nomiss(ncomp, flica_niter)
}
\arguments{
\item{ncomp}{number of components to be used in LICA}

\item{flica_niter}{number of iterations to be used in LICA}
}
\value{
The resulting dataset is the results estimated from LICA
}
\description{
This function calls the matlab function to run LICA on no-missing data
}
\examples{
# Save data at "./MATLAB_code/flica/data_nomiss.mat"
#writeMat("./MATLAB_code/flica/data_nomiss.mat",
#         mod1_true = mod1_true, mod2_true = mod2_true  # simulated data w/ no missing (before standardization, code in matlab will do the standardization)
#         )
# Run LICA on no-miss data
#re_nomiss = flica_nomiss(ncomp = 5, flica_niter = 1500)
}
