% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flica_replace0.R
\name{flica_replace0}
\alias{flica_replace0}
\title{Function to run LICA on data after replacing missing with 0}
\usage{
flica_replace0(ncomp, flica_niter)
}
\arguments{
\item{ncomp}{number of components to be used in LICA}

\item{flica_niter}{number of iterations to be used in LICA}
}
\value{
The resulting dataset is the results estimated from LICA
}
\description{
This function calls the matlab function to run LICA on data after replacing missing with 0
}
\examples{
# Save data at "./MATLAB_code/flica/data_replace0.mat"
#writeMat("./MATLAB_code/flica/data_replace0.mat",
#         mod1_miss = mod1_replace0, mod2_miss = mod2_replace0,      # data with missing as 0
#         mod1_std = mod1_replace0_std, mod2_std = mod2_replace0_std # data with missing as 0 after standardization
#         )
# Run current practice: replace NA with 0 + LICA
#re_replace0 = try(flica_replace0(ncomp = 5, flica_niter = 1500))
}
