% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LICA.R
\name{LICA}
\alias{LICA}
\title{Function to run LICA}
\usage{
LICA(
  data,
  ncomp = 5,
  niter = 1000,
  method = c("nomiss", "completer", "replace0")
)
}
\arguments{
\item{data}{data in lists}

\item{ncomp}{number of components to be used in LICA}

\item{niter}{number of iterations to be used in LICA}

\item{method}{LICA method ('nomiss','completer','replace0')}
}
\value{
The resulting dataset includes the results estimated from LICA
}
\description{
This function runs LICA for no-missing data and two current practices for missing data (complete case analysis and replaceing missing with 0s)
}
\examples{
# e.g., LICA on no-missing data
#LICA(data = data, ncomp = 5, niter = 1500, method = "nomiss")
}
