% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FI_LICA.R
\name{FI_LICA}
\alias{FI_LICA}
\title{Function to run FI-LICA}
\usage{
FI_LICA(
  data,
  ncomp = 5,
  flica_niter = 1500,
  n = 20,
  flica_niter2 = 1000,
  rescale = TRUE
)
}
\arguments{
\item{ncomp}{number of components to be used in LICA}

\item{flica_niter}{number of iterations to be used in FI-LICA's Step 1 (completer case analysis)}

\item{n}{number of FI-LICA updates}

\item{flica_niter2}{number of iterations to be used in FI-LICA's Step 2}

\item{rescale}{rescale H and XW (default is TRUE)}
}
\value{
The resulting dataset includes the results estimated from FI-LICA, convergence measure for H and for XW, and the dF history in LICA
}
\description{
This function runs FI-LICA. Note: current function only supports the analysis on 2 modalities.
}
\examples{
# generate data
data = data_generate_MAR(nsubj = 100, seed = 7452, n_miss = 5)
# data with missing
data = list(data$Y1_miss, data$Y2_miss)
# FI-LICA
#re_filica = FI_LICA(data = data, ncomp = 5, flica_niter = 1500, n = 20, flica_niter2 = 1000, rescale = TRUE)
}
