% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flica_step1.R
\name{filica_step1}
\alias{filica_step1}
\title{Function to run Step 1 in FI-LICA}
\usage{
filica_step1(re, rescale = TRUE)
}
\arguments{
\item{re}{the saved results from running LICA on completers}

\item{rescale}{rescale H and XW (default is TRUE)}
}
\value{
The resulting dataset includes the estimated H, XW, and the last dF from Step 1 in FI-LICA
}
\description{
This function runs Step 1 in FI-LICA. Note: current function only supports the analysis on 2 or 3 modalities.
}
\examples{
#re_step1 = filica_step1(re = re_completer, rescale = TRUE)
}
