% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flica_nomiss.R
\name{flica_nomiss}
\alias{flica_nomiss}
\title{Function to run LICA on no-missing data}
\usage{
flica_nomiss(ncomp, flica_niter)
}
\arguments{
\item{ncomp}{number of components to be used in LICA}

\item{flica_niter}{number of iterations to be used in LICA}
}
\value{
The resulting dataset is the results estimated from LICA
}
\description{
This function calls the matlab function to run LICA on no-missing data
}
